chore: move animation render function to lib
