+++ packages/client/src/core/index.ts
@@ -1,4 +1,5 @@
 export * from './game'
 export * from './hero'
+export * from './lib'
 export * from './maze'
 export * from './oranges'
