+++ packages/client/src/core/lib/random.ts
@@ -1,0 +1,5 @@
+export function random(a: number, b: number) {
+  return parseFloat(
+    (Math.random() * ((a ? a : 1) - (b ? b : 0)) + (b ? b : 0)).toFixed(1)
+  )
+}
